backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic>=2.7
chardet==5.2.0
python-pptx==0.6.23
reportlab==4.0.9
watchdog==4.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
PyMuPDF==1.23.26
pytest==8.0.0
pytest-asyncio==0.23.5
plyer==2.1.0

